requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq==0.3.8
langgraph
langchain-mcp-adapters
reportlab
requests
aiohttp
beautifulsoup4
lxml
pydantic
typing-extensions
python-docx
openai
httpx>=0.24.1
arxiv

# FastAPI Backend
fastapi
uvicorn[standard]

# Optional: Only needed if using local HuggingFace models
# langchain-huggingface
# transformers
# torch
# accelerate
# huggingface-hub
# MLOps Dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
sentence-transformers>=2.2.2
mlflow>=2.7.0
joblib>=1.3.0

dvc>=3.0.0
